+++ CHANGELOG.md
@@ -19,10 +19,10 @@
 - Improve story and documentation.
 - Mark components as `stable`.
 
-#### Tab navigation
+#### Tab navigation and disclosure button component
 
 - Add styles to demo application.
-- Improve documentation.
+- Improve story and documentation.
 - Mark component as `refining`.
 
 #### Icon and icon sprite component